Split test into tracers, add more tests
