+++ contracts/MultiplyBy7.sol
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: GPL-3.0-only
+pragma solidity >=0.8.2;
+
+contract MultiplyBy7 {
+    function multiply(uint256 a) public pure returns (uint256 d) {
+        return a * 7;
+    }
+}
